fix!: improve interoperability of SNYK_API environment variable
BREAKING CHANGE: The SNYK_API environment variable should be set to the same
value as intended for the Snyk CLI and no longer requires the "/rest" suffix.
Instead, only the base URL of the Snyk API shall be used.

To adapt to the new version, change the SNYK_API environment variable to omit
the "/rest", e.g. `SNYK_API="https://api.snyk.io`.